# PhotographyBack/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Mathis Petit",
        "description": "Documentation API pour le projet de photographie",
        "contact": {
            "email": "mathispetitfrpro@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "200": {
                        "description": "List of all categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "operationId": "71fcad552bb0eaba9fb191fd8d8dcab0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Landscape"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a specific category by ID",
                "operationId": "c68e76d323c008827a9e47398b1583de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update an existing category",
                "operationId": "0e686b2748211cc688333ed705dc111f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Portrait"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "operationId": "4c12e22a7f8c617bd83bfa1fdc05428d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
        "/api/partners/{id}": {
            "get": {
                "tags": [
                    "Partners"
                ],
                "summary": "Afficher un partenaire spécifique",
                "operationId": "3d2e93d00ae9df1efcbf6eecbf22496c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur associé au partenaire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations du partenaire",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Partner"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Partenaire non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Partners"
                ],
                "summary": "Mettre à jour un partenaire",
                "operationId": "183edc72e7cf428ad78060fbe8c13fce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur associé au partenaire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "bio": {
                                        "type": "string",
                                        "example": "Je suis un photographe passionné..."
                                    },
                                    "website": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "http://monsiteweb.com"
                                    },
                                    "social_links": {
                                        "type": "string",
                                        "example": "Instagram: @moncompte"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Informations mises à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Informations mises à jour avec succès"
                                        },
                                        "partner": {
                                            "$ref": "#/components/schemas/Partner"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Category model",
                "properties": {
                    "id": {
                        "description": "ID of the category",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the category",
                        "type": "string",
                        "example": "Landscape"
                    },
                    "created_at": {
                        "description": "Timestamp when the category was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-04T12:34:56.000000Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the category was updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-05T12:34:56.000000Z"
                    },
                    "photos": {
                        "description": "List of photos belonging to the category",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    }
                },
                "type": "object"
            },
            "Partner": {
                "title": "Partner",
                "description": "Modèle représentant un partenaire photographe dans l'application",
                "properties": {
                    "id": {
                        "description": "ID unique du partenaire",
                        "type": "integer"
                    },
                    "bio": {
                        "description": "Biographie du partenaire",
                        "type": "string"
                    },
                    "website": {
                        "description": "Site web du partenaire",
                        "type": "string"
                    },
                    "social_links": {
                        "description": "Liens vers les réseaux sociaux du partenaire",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur associé au partenaire",
                        "type": "integer"
                    },
                    "photos": {
                        "description": "Liste des photos associées à ce partenaire",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Photo": {
                "title": "Photo",
                "description": "Modèle représentant une photo dans l'application",
                "properties": {
                    "id": {
                        "description": "ID unique de la photo",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Titre de la photo",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description de la photo",
                        "type": "string"
                    },
                    "image_path": {
                        "description": "Chemin vers l'image de la photo",
                        "type": "string"
                    },
                    "likes": {
                        "description": "Nombre de likes que la photo a reçu",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur ayant téléchargé la photo",
                        "type": "integer"
                    },
                    "categories": {
                        "description": "Liste des catégories associées à cette photo",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "likedByUsers": {
                        "description": "Liste des utilisateurs ayant liké la photo",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Modèle représentant un utilisateur dans l'application",
                "properties": {
                    "id": {
                        "description": "ID unique de l'utilisateur",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom de l'utilisateur",
                        "type": "string"
                    },
                    "email": {
                        "description": "Adresse email de l'utilisateur",
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "description": "Rôle de l'utilisateur (user, partner, admin)",
                        "type": "string"
                    },
                    "profile_photo_url": {
                        "description": "URL de la photo de profil de l'utilisateur",
                        "type": "string"
                    },
                    "photos": {
                        "description": "Liste des photos associées à cet utilisateur",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    },
                    "partner": {
                        "$ref": "#/components/schemas/Partner"
                    },
                    "likedPhotos": {
                        "description": "Liste des photos likées par l'utilisateur",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}